requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
tqdm==4.66.1
python-dotenv==1.1.0
